combined appbar and sidebar into one component. Set trainer profile layout
