+++ my-app/src/components/Trainer/Dashboard/TrainerAppSidebar.js
@@ -18,26 +18,7 @@ import Link from '@material-ui/core/Link';
 import MenuIcon from '@material-ui/icons/Menu';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
 import NotificationsIcon from '@material-ui/icons/Notifications';
-// import { mainListItems, secondaryListItems } from './Dashboard/listItems';
-import TrainerAppSideBar from "./Dashboard/TrainerAppSidebar"
-import Students from "./Dashboard/Students";
-import Money from "./Dashboard/Money";
-import StudentTable from "./Dashboard/Table";
-import Chart from "./Dashboard/Chart";
-import StudentRequest from "./Dashboard/StudentRequests";
-
-function Copyright() {
-  return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {"Copyright © "}
-      <Link color="inherit" href="https://material-ui.com/">
-        Your Website
-      </Link>{" "}
-      {new Date().getFullYear()}
-      {"."}
-    </Typography>
-  );
-}
+import { mainListItems, secondaryListItems } from './listItems';
 
 const drawerWidth = 240;
 
@@ -120,17 +101,15 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-
-
-export default function TrainerDashboard() {
+export default function TrainerAppSideBar() {
   const classes = useStyles();
-  // const [open, setOpen] = React.useState(true);
-  // const handleDrawerOpen = () => {
-  //   setOpen(true);
-  // };
-  // const handleDrawerClose = () => {
-  //   setOpen(false);
-  // };
+  const [open, setOpen] = React.useState(true);
+  const handleDrawerOpen = () => {
+    setOpen(true);
+  };
+  const handleDrawerClose = () => {
+    setOpen(false);
+  };
   const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
 
   const cards = [1, 2, 3, 4, 5, 6];
@@ -140,48 +119,45 @@ export default function TrainerDashboard() {
 
       {/* <StudentTable /> */}
 
-      
 
-      <CssBaseline />
-      <TrainerAppSideBar />
 
-      {/* <AppBar
-      style={{
-        marginTop: "4%"
-      }}
-        position="absolute"
-        className={clsx(classes.appBar, open && classes.appBarShift)}
-      >
-        <Toolbar className={classes.toolbar}>
-          <IconButton
-            edge="start"
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            className={clsx(
-              classes.menuButton,
-              open && classes.menuButtonHidden
-            )}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Typography
-            component="h1"
-            variant="h6"
-            color="inherit"
-            noWrap
-            className={classes.title}
-          >
-            Trainer Dashboard
-          </Typography>
-          <IconButton color="inherit">
-            <Badge badgeContent={4} color="secondary">
-              <NotificationsIcon />
-            </Badge>
-          </IconButton>
-        </Toolbar>
-      </AppBar>
-      <Drawer
+  <AppBar
+  style={{
+    marginTop: "5%"
+  }}
+  position="absolute"
+  className={clsx(classes.appBar, open && classes.appBarShift)}
+  >
+  <Toolbar className={classes.toolbar}>
+    <IconButton
+      edge="start"
+      color="inherit"
+      aria-label="open drawer"
+      onClick={handleDrawerOpen}
+      className={clsx(
+        classes.menuButton,
+        open && classes.menuButtonHidden
+      )}
+    >
+      <MenuIcon />
+    </IconButton>
+    <Typography
+      component="h1"
+      variant="h6"
+      color="inherit"
+      noWrap
+      className={classes.title}
+    >
+      Trainer Dashboard
+    </Typography>
+    <IconButton color="inherit">
+      <Badge badgeContent={4} color="secondary">
+        <NotificationsIcon />
+      </Badge>
+    </IconButton>
+  </Toolbar>
+  </AppBar>
+  <Drawer
         variant="permanent"
         classes={{
           paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose),
@@ -197,43 +173,8 @@ export default function TrainerDashboard() {
         <List>{mainListItems}</List>
         <Divider />
         <List>{secondaryListItems}</List>
-      </Drawer> */}
-      <main className={classes.content}>
-        <div className={classes.appBarSpacer} />
-        <Container maxWidth="lg" className={classes.container}>
-          <Grid container spacing={3}>
-            {/* Chart */}
-            <Grid item xs={12} md={8} lg={9}>
-              <Paper className={fixedHeightPaper}>
-                <Chart />
-              </Paper>
-            </Grid>
-            {/* Recent Deposits */}
-            <Grid item xs={12} md={4} lg={3}>
-              <Paper className={fixedHeightPaper}>
-                <Money />
-              </Paper>
-            </Grid>
-            {/* Recent Orders */}
-            <Grid item xs={12}>
-              <Paper className={classes.paper}>
-                <StudentTable />
-              </Paper>
-            </Grid>
-          </Grid>
-          <Box pt={4}>
-            <Copyright />
-          </Box>
-          <Box pt={4}>
-            <Students cards={cards} />
-          </Box>
-          <Box pt={4}>
-          <StudentRequest />
-          </Box>
-        </Container>
-      </main>
-    </div>
+      </Drawer>
+</div>
 
-      
-  );
+)
 }
